refactor(#7): 멤버 API응답 분리
Member API response데이터 분리
