+++ Where_Are_You/Data/Models/Responses/GenericResponse.swift
@@ -12,26 +12,3 @@ struct GenericResponse<T: Codable>: Codable {
     let message: String
     let data: T
 }
-
-struct LoginResponse: Codable {
-    let accessToken: String
-    let refreshToken: String
-    let memberSeq: Int
-    let memberCode: String
-}
-
-struct MemberSearchResponse: Codable {
-    let userName: String
-    let memberSeq: Int
-    let profileImage: String
-}
-
-struct MemberDetailsResponse: Codable {
-    let userName: String
-    let email: String
-    let profileImage: String
-}
-
-struct CheckEmailResponse: Codable {
-    let email: String
-}
