+++ Where_Are_You/Data/Models/Responses/Member/CheckEmailResponse.swift
@@ -1,0 +1,12 @@
+//
+//  CheckEmailResponse.swift
+//  Where_Are_You
+//
+//  Created by 오정석 on 5/8/2024.
+//
+
+import Foundation
+
+struct CheckEmailResponse: Codable {
+    let email: String
+}
