+++ Where_Are_You/Data/Models/Responses/Member/MemberDetailsResponse.swift
@@ -1,0 +1,14 @@
+//
+//  MemberDetailsResponse.swift
+//  Where_Are_You
+//
+//  Created by 오정석 on 5/8/2024.
+//
+
+import Foundation
+
+struct MemberDetailsResponse: Codable {
+    let userName: String
+    let email: String
+    let profileImage: String
+}
